requirements-dev.txt
pytest>=8.3
pyyaml>=6.0
